requirements.txt
fastapi
uvicorn
requests
authlib
psycopg2-binary
httpx
sqlalchemy
itsdangerous
python-dotenv
